cmake: use ${CMAKE_DL_LIBS} instead of dl
Most platforms do not require linking against libdl, but NetBSD requires
explicitly linking with libdl. Therefore, it is incorrect to hardcode a
requirement on this library. CMake provides the variable CMAKE_DL_LIBS that is
intended to contain the platform-appropriate libraries. The following patch
replaces the use of 'dl' with the variable.

Needed for #32